python_telegram_bot/requirements.txt
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.31.0
python-telegram-bot-pagination==0.0.2
python-telegram-bot-calendar==1.0.5
loguru==0.6.0